Move popperjs mock to jest setup file
