+++ .storybook/main.js
@@ -15,7 +15,13 @@ module.exports = {
 			},
 		},
 	],
-	stories: ['../README.story.mdx', '../docs/**/*.story.mdx', '../src/**/*.story.jsx'],
+	stories: [
+		'../README.story.mdx',
+		'../docs/**/*.story.mdx',
+		'../src/**/*.story.jsx',
+		'../docs/**/*.stories.mdx',
+		'../src/**/*.stories.jsx',
+	],
 	webpackFinal: async (config, { configType }) => {
 		const { hasFoundAny, matches } = getLoaders(config, loaderByName('babel-loader'))
