+++ src/components/Constraints/stories/Popup.stories.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+import { Organism } from '../Organism'
+
+export default {
+	title: 'Components/Constraint',
+	parameters: {
+		componentSubtitle: 'Popup Cards for Constraints',
+	},
+}
